Handle `Fill.min` bounded constraints in `flex` layout
