+++ Cargo.toml
@@ -120,6 +120,9 @@ criterion = "0.5"
 iced_wgpu.workspace = true
 iced_renderer.workspace = true
 
+iced.workspace = true
+iced.features = ["advanced"]
+
 [[bench]]
 name = "wgpu"
 harness = false
